refactorizing module auth
