+++ src/css/signin.css
@@ -7,6 +7,7 @@
   text-align: center;
   width: 100%;
   box-sizing: border-box;
+  padding: 35px;
 }
 
 .img-home {
@@ -23,7 +24,7 @@
   display: grid;
   width: 100%;
   max-width: 400px;
-  min-width: 250px ;
+  min-width: 250px;
   height: 70vh;
   justify-items: center;
   align-items: center;
@@ -34,14 +35,15 @@
 
 .email-signin, .password-signin, .submit-signin {
   width: 75%;
-  height: 38px;
+  height: 40px;
   border: none;
   border-radius: 20px;
   padding: 5px 15px;
-  margin: 10px;
   box-sizing: border-box;
-  box-shadow: 0px 4px 40px rgba(147, 147, 147, 0.4);
-  background-color: #ffffff;
+  border: 1px solid #DDDDDD;
+  border: 0;
+  box-shadow: 1px 1px 1px 1px rgba(0, 0, 0, 0.15);
+  font-family: 'Nunito Sans', sans-serif;
 }
 
 #signin {
@@ -60,6 +62,7 @@
 .change-form {
   color: #FF009E;
   font-weight: bold;
+  padding-left: 10px;
 }
 
 .text-signin a {
