--- src/firebase-controller/signout-controller.js
@@ -1,6 +1,0 @@
-import { signOut } from '../firebase/autentication.js';
-
-export const signOutUser = () => signOut()
-  .then(() => {
-    window.location.hash = '#/signin';
-  });
