+++ src/view/signin.js
@@ -1,5 +1,4 @@
-import { signinUser } from '../firebase-controller/signin-controller.js';
-import { googleAccount } from '../firebase-controller/authGoogle.js';
+import * as auth from '../auth/index.js';
 
 export default () => {
   const viewSignIn = `
@@ -7,16 +6,16 @@ export default () => {
   <form id="signin-form">
     <img class="logo-signin" src="./img/logo-lab-black.svg" alt="">
     <p class="text-signin">Una red social para preguntas y respuestas hecho por laboratorians y para laboratorians.</p>
-    <button type="button" class="button-google-signin"><img src="./img/logo-google.svg" class="google-icon"alt="">Google</button>
+    <input class ="email-signin" type="email" id="mail" name="user_mail" placeholder="Ingresa tu correo electrónico">
+    <input class = "password-signin" type="password" id="passwordSignIn" name="user_password" placeholder="Ingresa tu contraseña">
+    <div id= "errorMessage"></div>
+    <input class="submit-signin" type="submit" id="signin" value="Iniciar Sesión">
     <div class="line">
       <span class="text-signin">
         O continua con
       </span>
     </div>
-    <input class ="email-signin" type="email" id="mail" name="user_mail" placeholder="Ingresa tu correo electrónico">
-    <input class = "password-signin" type="password" id="passwordSignIn" name="user_password" placeholder="Ingresa tu contraseña">
-    <div id= "errorMessages"></div>
-    <input class="submit-signin" type="submit" id="signin" value="Iniciar Sesión">
+    <img src="./img/logo-google.svg" class="google-icon"alt="">
     <p class="text-signin">¿No tienes una cuenta? <a href="#/"><span class="change-form">Regístrate</span></a></p>
   </form>`;
   document.getElementById('container').classList.add('main');
@@ -27,12 +26,11 @@ export default () => {
   const form = sectionElement.querySelector('#signin-form');
   const passwordField = sectionElement.querySelector('#passwordSignIn');
   const emailField = sectionElement.querySelector('#mail');
-  const google = sectionElement.querySelector('.button-google-signin');
-  //   Fields validation
+  const google = sectionElement.querySelector('.google-icon');
   form.addEventListener('submit', (e) => {
     e.preventDefault();
-    signinUser(emailField.value, passwordField.value);
+    auth.signinUser(emailField.value, passwordField.value);
   });
-  google.addEventListener('click', () => googleAccount());
+  google.addEventListener('click', () => auth.googleAccount());
   return sectionElement;
 };
